refactor: packages로 공통 유틸/훅 이동
