+++ apps/admin/package.json
@@ -15,6 +15,8 @@
   },
   "dependencies": {
     "@surf/ui": "workspace:*",
+    "@surf/hooks": "workspace:*",
+    "@surf/utils": "workspace:*",
     "@tanstack/react-query": "^5.90.16",
     "@tanstack/react-query-devtools": "^5.91.2",
     "axios": "^1.12.2",
