+++ apps/web/src/app-pages/post/PostDetailPage.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { useKeyboardOffset } from '@surf/hooks';
 import { HeaderMode } from '@surf/ui/header';
 import { useAlertStore } from '@surf/ui/store/alertStore';
 import { useToastStore } from '@surf/ui/store/toastStore';
@@ -12,7 +13,6 @@ import { useAuthStore } from '@/features/auth/model/useAuthStore';
 import { useDeletePostMutation } from '@/features/post/model/useDeletePostMutation';
 import { useGetPostLikesQuery } from '@/features/post/model/useGetPostLikesQuery';
 import { useGetPostScheduleQuery } from '@/features/post/model/useGetPostScheduleQuery';
-import { useKeyboardOffset } from '@/shared/hooks/useKeyboardOffset';
 import { useBottomSheetStore } from '@/shared/store/bottomSheetStore';
 import { CommentComposer } from '@/widgets/comment-composer/ui/CommentComposer';
 import { CommentSection } from '@/widgets/comment-section/ui/CommentSection';
