+++ apps/web/src/entities/notification/ui/NotificationItem.tsx
@@ -1,7 +1,7 @@
 import { Avatar } from '@surf/ui/avatar';
+import { formatDateTime } from '@surf/utils';
 import LikeNotificationBadge from '@/shared/assets/icons/notification/notification-badge-like.svg';
 import MentionNotificationBadge from '@/shared/assets/icons/notification/notification-badge-mention.svg';
-import { formatDateTime } from '@/shared/utils/date';
 
 export type NOTIFICATION_BADGE = 'LIKE' | 'MENTION';
