+++ apps/web/src/features/image/model/useImageSelector.ts
@@ -1,9 +1,8 @@
 'use client';
 
+import { safeUUID, reorderArray } from '@surf/utils';
 import { useEffect, useRef, useState } from 'react';
-import { reorderArray } from '@/shared/utils/reorder';
 import { UploadImage } from '@/entities/image/model/types';
-import { safeUUID } from '@/shared/utils/uuid';
 
 /**
  * 이미지 선택 및 로컬 미리보기, 순서 변경을 담당하는 훅.
