--- apps/web/src/shared/utils/date.ts
@@ -1,42 +1,0 @@
-export function toDate(isoString: string): Date {
-  return new Date(isoString);
-}
-
-export function toKST(date: Date) {
-  const utc = date.getTime() + date.getTimezoneOffset() * 60000;
-  return new Date(utc + 9 * 60 * 60 * 1000);
-}
-
-export function formatDate(date: Date) {
-  const y = date.getFullYear();
-  const m = String(date.getMonth() + 1).padStart(2, '0');
-  const d = String(date.getDate()).padStart(2, '0');
-  return `${y}.${m}.${d}`;
-}
-
-export function formatDateTime(date: Date) {
-  const y = date.getFullYear();
-  const m = String(date.getMonth() + 1).padStart(2, '0');
-  const d = String(date.getDate()).padStart(2, '0');
-  const hh = String(date.getHours()).padStart(2, '0');
-  const mm = String(date.getMinutes()).padStart(2, '0');
-  return `${y}.${m}.${d} ${hh}:${mm}`;
-}
-
-export function timeAgo(date: Date): string {
-  const now = new Date();
-  const diff = (now.getTime() - date.getTime()) / 1000;
-
-  if (diff < 60) return '방금 전';
-  if (diff < 3600) return `${Math.floor(diff / 60)}분 전`;
-  if (diff < 86400) return `${Math.floor(diff / 3600)}시간 전`;
-
-  return formatDate(date);
-}
-
-// Date 객체를 받아 MM.DD로 변환
-export function formatMonthDay(date: Date) {
-  const m = String(date.getMonth() + 1).padStart(2, '0');
-  const d = String(date.getDate()).padStart(2, '0');
-  return `${m}.${d}`;
-}
