+++ apps/web/src/widgets/profile-badge/ui/ProfileBadge.tsx
@@ -1,12 +1,12 @@
 'use client';
 
+import { useInfiniteScroll } from '@surf/hooks';
 import { ActivityBadge } from '@surf/ui/activity-badge';
 // import dynamic from 'next/dynamic';
 import { useMemo, useRef } from 'react';
 import { dedupeAndSortBadges } from '../model/utils';
 import { useInfiniteBadges } from '@/entities/user/model/useInfiniteBadges';
 import ActivityBadgeEmpty from '@/shared/assets/icons/empty-space/activity-badge-empty.svg';
-import { useInfiniteScroll } from '@/shared/hooks/useInfiniteScroll';
 
 // const ActivityBadgeEmpty = dynamic(
 //   () => import('@/shared/assets/icons/empty-space/activity-badge-empty.svg'),
