+++ eslint.config.mjs
@@ -1,4 +1,6 @@
 import web from './apps/web/eslint.config.js';
 import ui from './packages/ui/eslint.config.js';
+import hooks from './packages/hooks/eslint.config.js';
+import utils from './packages/utils/eslint.config.js';
 
-export default [...web, ...ui];
+export default [...web, ...ui, ...hooks, ...utils];
